pyproject.toml
[tool.poetry]
name = "codeforcesparser"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.31"
requests = "^2.32.3"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
aiogram = "^3.10.0"
asyncpg = "^0.29.0"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
aiogram-tests = "^1.0.3"
pytest-asyncio = "^0.23.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
